+++ src/types/date-time.ts
@@ -1,7 +1,5 @@
-import type { Branded } from './branded.js';
+export type DateInput = ConstructorParameters<typeof Date>[0];
 
 export type DateString = `${number}-${number}-${number}`;
 
 export type TimeString = `${number}:${number}:${number}`;
-
-export type ISODateString = Branded<`${DateString}T${TimeString}.${number}Z`, 'ISODateString'>;
