+++ src/types/functions.ts
@@ -1,0 +1,9 @@
+export type Predicate<A, T extends A = A> =
+  | ((value: T, index: number, array: readonly A[]) => value is T)
+  | ((value: T, index: number, array: readonly A[]) => unknown);
+
+export type TypePredicateFn<T> = (input: unknown) => input is T;
+
+export type TypeAssertionFn<T> = (input: unknown) => asserts input is T;
+
+export type CompareFn<T> = (a: T, b: T) => number;
